Lots of refactors
