+++ frontend/src/components/Search/OverlayButton.module.css
@@ -8,6 +8,6 @@
   color: white;
   border: none;
   cursor: pointer;
-  box-shadow: 2px 2px 10px rgba(0 0 0 0.2);
-  z-index: 2;
+  box-shadow: 2px 2px 10px rgb(0 0 0 / 20%);
+  z-index: 1200;
 }
