+++ frontend/src/components/Search/OverlayComponent.module.css
@@ -1,0 +1,41 @@
+.overlayModal {
+  position: fixed;
+  bottom: 0;
+  width: 43.75%;
+  height: 80%;
+  background: var(--color-surface);
+  transition: right 0.5s ease-in-out;
+  z-index: 1100;
+  border: 2px solid var(--color-border);
+  box-shadow: 0 0 10px 10px rgb(0 0 0 / 20%);
+  overflow-y: auto;
+  padding: 10px;
+  border-radius: 10px;
+}
+
+.overlayModalHidden {
+  right: -100%;
+}
+
+.overlayModalVisible {
+  right: 0;
+}
+
+.expandBtn {
+  padding: 3px;
+  border-radius: 6px;
+  cursor: pointer;
+  background-color: var(--color-select-multivalue);
+  color: var(--color-text-secondary);
+  position: absolute;
+  top: 0%;
+  z-index: 1101;
+  transition:
+    transform 0.05s linear,
+    background-color var(--trans-dur),
+    color var(--trans-dur);
+}
+
+.expandIcon {
+  display: block;
+}
