--- frontend/src/pages/OverlayComponent.module.css
@@ -1,20 +1,0 @@
-.overlayModal {
-  position: fixed;
-  width: 43.75%;
-  height: 100%;
-  background-color: white;
-  transition: right 0.5s ease-in-out;
-  z-index: 1;
-  box-shadow: -2px 0 8px rgba(0 0 0 20%);
-  overflow-y: auto;
-  padding: 10px;
-  border-radius: 10px;
-}
-
-.overlayModalHidden {
-  right: -100%;
-}
-
-.overlayModalVisible {
-  right: 0;
-}
